Permission
A vendor cannot delete or update a product that doesn't belong to him, but the admin can.

+++ routes/api.php
@@ -20,19 +20,15 @@ use Illuminate\Support\Facades\Route;
 Route::group(['middleware'=>'auth:sanctum'], function(){
 
     Route::group(['controller'=>CategoryController::class], function(){
-        Route::get('/categories', 'index');
-        Route::post('/categories', 'store');
-        Route::get('/categories/{id}', 'show');
-        Route::put('/categories/{id}', 'update');
-        Route::delete('/categories/{id}', 'destroy');
+        Route::post('/categories', 'store')->middleware('permission:store category');
+        Route::put('/categories/{id}', 'update')->middleware('permission:update category');
+        Route::delete('/categories/{id}', 'destroy')->middleware('permission:delete category');
     });
 
     Route::group(['controller' => ProductController::class], function(){
-        Route::get('/products', 'index');
-        Route::post('/products', 'store');
-        Route::get('/products/{id}', 'show');
-        Route::put('/products/{id}', 'update');
-        Route::delete('/products/{id}', 'destroy');
+        Route::post('/products', 'store')->middleware('permission:store product');
+        Route::put('/products/{id}', 'update')->middleware('permission:update product|update every product');
+        Route::delete('/products/{id}', 'destroy')->middleware('permission:delete my product|delete every product');
     });
 
     Route::put('/profile/{user}', [AuthenticationController::class, 'update']);
@@ -40,3 +36,8 @@ Route::group(['middleware'=>'auth:sanctum'], function(){
 
 Route::post('/register', [AuthenticationController::class, 'register']);
 Route::post('/login', [AuthenticationController::class, 'login']);
+
+Route::get('/categories', [CategoryController::class,'index']);
+Route::get('/categories/{id}', [CategoryController::class,'show']);
+Route::get('/products', [ProductController::class, 'index']);
+Route::get('/products/{id}', [ProductController::class, 'show']);
